Fixed: USB to serial ZLP problem.
When using the Upgrade Firmware Utility to load a firmware into FA 300 panels. A sequence of 512 bytes was send by the program to the USB converter followed by a pause, this caused a zero length package to be sent. The double buffer reception on the USB FS driver was ignoring such events and causing the driver to lock.